// src/app/models/immigrationviewclientprofile.ts
export class ImmigrationViewClientProfile {
    fileNumber: string;
    status: string;
    creationDate: string;
    employeeId: string;
    email: string;
    firstName: string;
    middleName: string;
    lastName: string;
    phoneNumber: string;
    markForDeletion: string;
    deletedByUserOn: string;
    lastUpdatedByUserOn:string;
    createdByUserOn:string;
}